Delete unnecessary `GettingStartedArticleRenderer` component
